lib/software/mqtt_client/requirements_frozen.txt
appdirs==1.4.3
packaging==16.8
paho-mqtt==1.1
pyparsing==2.2.0
six==1.10.0
